requirements.txt
crewai>=0.28.1
pyyaml>=6.0
ollama>=0.1.5
langchain>=0.0.267
python-dotenv>=1.0.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.97.0
nltk>=3.8.1
scikit-learn>=1.3.0 
crewai-tools
requests